chore: remove compatibility layers for Symfony 3 and 4
